+++ bbr/Commands/Connect.cs
@@ -10,7 +10,7 @@ namespace bbr.Commands
     public class Connect : Command
     {
         public const byte COMMAND_ID = 1;
-        public override int CommandId => COMMAND_ID;
+        public override byte CommandId => COMMAND_ID;
 
         public int ConnectionId { get; protected set; }
