+++ bbr/Commands/Purge.cs
@@ -9,8 +9,8 @@ namespace bbr.Commands
 {
     public class Purge : Command
     {
-        public const int COMMAND_ID = 3;
-        public override int CommandId => COMMAND_ID;
+        public const byte COMMAND_ID = 3;
+        public override byte CommandId => COMMAND_ID;
 
         public int ConnectionId { get; protected set; }
